Feature/deploy

+++ server/src/seeds/cleanDb.ts
@@ -2,15 +2,14 @@ import models from '../models/index.js';
 import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
-  try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+  if (!models[modelName].db.db) {
+    throw new Error(`Model or database for ${modelName} does not exist`);
+  }
+  const modelExists = await models[modelName].db.db.listCollections({
+    name: collectionName
+  }).toArray();
 
-    if (modelExists.length) {
-      await db.dropCollection(collectionName);
-    }
-  } catch (err) {
-    throw err;
+  if (modelExists.length) {
+    await db.dropCollection(collectionName);
   }
 }
